MArginal ePIstasis Test for Regions (MAPITR)

Run MAPITR for a group of pathways on a single phenotype and a set of genome-wide SNPs

MAPITR(
  Genotypes,
  Phenotype,
  Pathways,
  Covariates = NULL,
  CenterStandardize = TRUE,
  ...
)

Arguments

Genotypes
A n x p matrix containing the genotypes (0/1/2) for all p SNPs across all n individuals. No default value.

Phenotype
A vector containing phenotypic values for all individuals being analyzed. No default value.

Pathways
A r x 2 matrix containing the pathway names and then a comma-separated list of the Genotypes column indices representing each SNP in the associated pathway. Note, this second column of comma-separated indices are the numeric positions for each SNP in Genotypes and not the SNP IDs or column names. No default value.

Covariates
A n x q matrix containing any q additional covariates that should be included in the M-projection matrix of the model. See Turchin et al. 2020 for details. Note that these are covariates which are applied to both sides of the model, ie the phenotype as well as the genotypes. A y-intercept term is automatically included and does not need to be part of this n x q matrix. No default value.

CenterStandardize
A logical TRUE/FALSE flag that indicates whether the genotype matrix Genotypes should be centered and standardized before analysis. This is a recommended step. Indicate FALSE if this is a preprocessing step that has already been done prior to running MAPITR. The default value is TRUE.

...
Additional optional arguments.

Value

A matrix containing in the first column the list of pathways that were analyzed, in the second column the associated MAPITR p-values for each pathway, and in the third column the associated MAPITR PVEs for each pathway.

Examples

data(MAPITR_TestData_Genotypes, MAPITR_TestData_Phenotype,
MAPITR_TestData_Pathways)
MAPITROutput <- MAPITR(MAPITR_TestData_Genotypes, MAPITR_TestData_Phenotype,
MAPITR_TestData_Pathways)
MAPITROutput$Results

#>      Pathways    pValues        Est        PVE
#> [1,]        1 0.00511314 1.09663702 1.32243728
#> [2,]        2 0.31118083 0.40971253 0.45740085
#> [3,]        3 0.33131510 0.36199437 0.43933281
#> [4,]        4 0.96134690 0.01706105 0.02017861
#> [5,]        5 0.94181499 0.02552678 0.03176702